+++ examples/openai/omni_scraper_openai.py
@@ -19,7 +19,7 @@ openai_key = os.getenv("OPENAI_APIKEY")
 graph_config = {
     "llm": {
         "api_key": openai_key,
-        "model": "gpt-4-turbo",
+        "model": "gpt-4o",
     },
     "verbose": True,
     "headless": True,