+++ scrapegraphai/nodes/text_to_speech_node.py
@@ -3,6 +3,8 @@ TextToSpeechNode Module
 """
 
 from typing import List, Optional
+
+from ..utils.logging import get_logger
 from .base_node import BaseNode
 
 
@@ -21,12 +23,19 @@ class TextToSpeechNode(BaseNode):
         node_name (str): The unique identifier name for the node, defaulting to "TextToSpeech".
     """
 
-    def __init__(self, input: str, output: List[str],
-                 node_config: Optional[dict]=None, node_name: str = "TextToSpeech"):
+    def __init__(
+        self,
+        input: str,
+        output: List[str],
+        node_config: Optional[dict] = None,
+        node_name: str = "TextToSpeech",
+    ):
         super().__init__(node_name, "node", input, output, 1, node_config)
 
         self.tts_model = node_config["tts_model"]
-        self.verbose = False if node_config is None else node_config.get("verbose", False)
+        self.verbose = (
+            False if node_config is None else node_config.get("verbose", False)
+        )
 
     def execute(self, state: dict) -> dict:
         """
@@ -35,7 +44,7 @@ class TextToSpeechNode(BaseNode):
         Args:
             state (dict): The current state of the graph. The input keys will be used to fetch the
                             correct data types from the state.
-                            
+
         Returns:
             dict: The updated state with the output key containing the audio generated from the text.
 
@@ -44,8 +53,7 @@ class TextToSpeechNode(BaseNode):
                         necessary information for generating the audio is missing.
         """
 
-        if self.verbose:
-            print(f"--- Executing {self.node_name} Node ---")
+        self.logger.info(f"--- Executing {self.node_name} Node ---")
 
         # Interpret input keys based on the provided input expression
         input_keys = self.get_input_keys(state)